Create Rainbow Renzene
Turtle is a special feathers of Python. Using Turtle, we can easily draw in a drawing board.

First we import the turtle module. Then create a window, next we create turtle object and using turtle method we can draw in the drawing board.

We can create so many shapes like a square , star, various patterns of square and so on. Also, with turtle programming , we can create very unique patterns using very simple code as I have done in this file.